Add files via upload
This project work helped me understand how financial institutions make decisions on lending approvals. So far from working on such project I have come to understanding why it is important to understand the algorithms and the math behind the algorithm as this will play a key role in having a complete understanding of the end to end process.